+++ network/src/main/java/com/nexters/misik/network/dto/GetReviewResponseDto.kt
@@ -1,6 +1,5 @@
-package com.nexters.misik.data.dto
+package com.nexters.misik.network.dto
 
-import com.nexters.misik.domain.ReviewEntity
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
@@ -13,10 +12,4 @@ data class GetReviewResponseDto(
     @SerialName("review")
     val review: String?,
 
-) {
-    fun toDomain() = ReviewEntity(
-        id = id,
-        isSuccess = isSuccess,
-        review = review,
-    )
-}
+)
