+++ network/src/main/java/com/nexters/misik/network/service/ReviewService.kt
@@ -1,7 +1,7 @@
-package com.nexters.misik.data.service
+package com.nexters.misik.network.service
 
-import com.nexters.misik.data.dto.GenerateReviewRequestDto
-import com.nexters.misik.data.dto.GetReviewResponseDto
+import com.nexters.misik.network.dto.GenerateReviewRequestDto
+import com.nexters.misik.network.dto.GetReviewResponseDto
 import retrofit2.http.Body
 import retrofit2.http.GET
 import retrofit2.http.POST
